cleaned api & started to create monitoring (/stats endpoint)
