--- log copy.txt
@@ -1,130 +1,0 @@
-rxm@Pro ts % yarn start
-yarn run v1.22.22
-$ tsx src/index.ts
-Wallet address 0x267094BA2E56487729F8406F6A5cD50cCEd2359D
-getAccountJson {
-  "success": true,
-  "data": {
-    "user_id": 230561,
-    "account_id": "0x9ab70de44988c824b82890cf4685862bfd3a58cc551737b73fbc46d362f821f6"
-  },
-  "timestamp": 1753377840224
-}
-getClientHolding: {
-  "success": true,
-  "data": {
-    "holding": []
-  },
-  "timestamp": 1753377840538
-}
-getOrderbook: {
-  "success": true,
-  "data": {
-    "timestamp": 1753377837538,
-    "asks": [
-      {
-        "price": 3731.51,
-        "quantity": 18.2936
-      }
-    ],
-    "bids": [
-      {
-        "price": 3712.71,
-        "quantity": 39.9503
-      }
-    ]
-  },
-  "timestamp": 1753377840923
-}
-creating order {
-  "symbol": "PERP_ETH_USDC",
-  "order_type": "LIMIT",
-  "side": "BUY",
-  "order_price": 3638,
-  "order_quantity": 0.01
-}
-createOrder: {
-  "success": false,
-  "code": -1101,
-  "message": "The margin will be insufficient after."
-}
-getOpenOrders: {
-  "success": true,
-  "data": {
-    "rows": [],
-    "meta": {
-      "total": 0,
-      "records_per_page": 25,
-      "current_page": 1
-    }
-  },
-  "timestamp": 1753377841766
-}
-creating algo order {
-  "symbol": "PERP_ETH_USDC",
-  "algo_type": "STOP",
-  "quantity": 0.01,
-  "trigger_price": 4084,
-  "trigger_price_type": "MARK_PRICE",
-  "side": "SELL",
-  "type": "MARKET",
-  "reduce_only": true
-}
-createAlgoOrder: {
-  "success": true,
-  "data": {
-    "rows": [
-      {
-        "order_id": 49605761,
-        "client_order_id": null,
-        "algo_type": "TAKE_PROFIT",
-        "order_quantity": 0.01
-      }
-    ]
-  },
-  "timestamp": 1753377842050
-}
-getOpenAlgoOrders: {
-  "success": true,
-  "data": {
-    "rows": [
-      {
-        "algo_order_id": 49605761,
-        "root_algo_order_id": 49605761,
-        "parent_algo_order_id": 0,
-        "symbol": "PERP_ETH_USDC",
-        "algo_type": "TAKE_PROFIT",
-        "side": "SELL",
-        "quantity": 0.01,
-        "is_triggered": false,
-        "trigger_price": 4084,
-        "trigger_status": "USELESS",
-        "type": "MARKET",
-        "root_algo_status": "NEW",
-        "algo_status": "NEW",
-        "trigger_price_type": "MARK_PRICE",
-        "trigger_time": 0,
-        "total_executed_quantity": 0,
-        "average_executed_price": 0,
-        "total_fee": 0,
-        "fee_asset": "USDC",
-        "reduce_only": true,
-        "created_time": 1753377842025,
-        "updated_time": 1753377842035,
-        "realized_pnl": 0
-      }
-    ],
-    "meta": {
-      "total": 1,
-      "records_per_page": 25,
-      "current_page": 1
-    }
-  },
-  "timestamp": 1753377842331
-}
-cancelAlgoOrder: {
-  "success": true,
-  "status": "CANCEL_SENT"
-}
-✨  Done in 4.01s.
-rxm@Pro ts %
