--- package copy.json
@@ -1,40 +1,0 @@
-{
-  "private": true,
-  "name": "orderly-v2-examples-ts",
-  "version": "1.0.0",
-  "description": "Examples of how to use Orderly EVM API",
-  "author": "Mario Reder <mario@orderly.network>",
-  "license": "MIT",
-  "engines": {
-    "node": ">=20"
-  },
-  "scripts": {
-    "start": "tsx src/index.ts",
-    "typegen": "typechain --target=ethers-v6 --input-dir ./src/abi --out-dir ./src/abi",
-    "lint": "prettier --check . && eslint . --ext .js,.jsx,.ts,.tsx,.svelte",
-    "format": "prettier --write . && yarn eslint --fix ."
-  },
-  "dependencies": {
-    "@noble/ed25519": "^2.0.0",
-    "@solana/web3.js": "^1.98.0",
-    "bs58": "^5.0.0",
-    "dotenv": "^16.4.5",
-    "ethereum-cryptography": "^3.1.0",
-    "ethers": "^6.13.5"
-  },
-  "devDependencies": {
-    "@orderly.network/types": "^0.2.4",
-    "@typechain/ethers-v6": "^0.5.1",
-    "@typescript-eslint/eslint-plugin": "^7.2.0",
-    "@typescript-eslint/parser": "^7.2.0",
-    "eslint": "^8.57.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-config-typescript": "^3.0.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-prettier": "^5.1.3",
-    "prettier": "^3.2.5",
-    "tsx": "^4.7.1",
-    "typechain": "^8.3.2",
-    "typescript": "^5.4.2"
-  }
-}
