+++ package.json
@@ -20,12 +20,13 @@
   "dependencies": {
     "@noble/ed25519": "^2.0.0",
     "@solana/web3.js": "^1.98.0",
+    "axios": "^1.11.0",
     "bs58": "^5.0.0",
+    "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "ethereum-cryptography": "^3.1.0",
     "ethers": "^6.13.5",
-    "express": "^4.18.2",
-    "cors": "^2.8.5"
+    "express": "^4.18.2"
   },
   "devDependencies": {
     "@orderly.network/types": "^0.2.4",
