--- seed copy.txt
@@ -1,47 +1,0 @@
-1.
-
-end
-
-2.
-
-merit
-
-3.
-
-excess
-
-4.
-
-street
-
-5.
-
-cost
-
-6.
-
-pear
-
-7.
-
-kite
-
-8.
-
-galaxy
-
-9.
-
-essay
-
-10.
-
-excite
-
-11.
-
-salmon
-
-12.
-
-stamp
